refactor(rule): rename cross-validation files to follow naming conventions
Use .stubs prefix for test helpers and .comparators suffix for comparator files.
