+++ libs/methodologies/bold/rule-processors/mass-id/document-manifest-data/src/cross-validation/cross-validation.stubs.ts
