+++ libs/methodologies/bold/rule-processors/mass-id/document-manifest-data/src/cross-validation/document.comparators.ts
@@ -9,8 +9,8 @@ import type {
   MtrNumbersComparison,
   ProcessingPeriodComparison,
 } from '../document-manifest-data.result-content.types';
-import type { ComparisonOutput } from './cross-validation.field-comparators';
 import type { FieldValidationResult } from './cross-validation.helpers';
+import type { ComparisonOutput } from './field.comparators';
 
 import { parsePeriodRange } from './cross-validation.helpers';
