+++ libs/methodologies/bold/rule-processors/mass-id/document-manifest-data/src/cross-validation/index.ts
@@ -1,4 +1,4 @@
-export * from './cross-validation.document-comparators';
-export * from './cross-validation.field-comparators';
 export * from './cross-validation.helpers';
-export * from './cross-validation.waste-comparators';
+export * from './document.comparators';
+export * from './field.comparators';
+export * from './waste.comparators';
