+++ libs/methodologies/bold/rule-processors/mass-id/document-manifest-data/src/cross-validation/waste.comparators.spec.ts
@@ -7,7 +7,7 @@ import {
   compareWasteQuantity,
   compareWasteType,
   WEIGHT_DISCREPANCY_THRESHOLD,
-} from './cross-validation.waste-comparators';
+} from './waste.comparators';
 
 const notExtractedReason: ReviewReason = {
   code: 'NOT_EXTRACTED',
