+++ libs/methodologies/bold/rule-processors/mass-id/document-manifest-data/src/recycling-manifest-cross-validation.helpers.spec.ts
@@ -8,7 +8,7 @@ import type { DocumentEvent } from '@carrot-fndn/shared/methodologies/bold/types
 import {
   stubEntity,
   stubEntityWithAddress,
-} from './cross-validation/cross-validation.test-helpers';
+} from './cross-validation/cross-validation.stubs';
 import {
   type CdfCrossValidationEventData,
   collectMtrDocumentNumbers,
